# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/karyawan/gaji": {
            "get": {
                "tags": [
                    "GajiKaryawan"
                ],
                "summary": "Ambil data gaji karyawan, dengan pencarian berdasarkan nama karyawan (wajib)",
                "operationId": "GetGaji",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci pencarian nama karyawan (wajib diisi)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data gaji karyawan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data gaji karyawan tersedia"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "nip": {
                                                        "type": "string",
                                                        "example": "6535534"
                                                    },
                                                    "periode": {
                                                        "type": "string",
                                                        "example": "05/2025"
                                                    },
                                                    "gaji_pokok": {
                                                        "type": "integer",
                                                        "example": 5000000
                                                    },
                                                    "potongan": {
                                                        "type": "integer",
                                                        "example": 500000
                                                    },
                                                    "total_gaji": {
                                                        "type": "integer",
                                                        "example": 4500000
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-06-02T10:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-06-02T11:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/karyawan/gaji/create": {
            "post": {
                "tags": [
                    "GajiKaryawan"
                ],
                "summary": "Tambah data gaji karyawan baru",
                "operationId": "CreateGajiKaryawan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_karyawan",
                                    "periode",
                                    "gaji_pokok",
                                    "potongan"
                                ],
                                "properties": {
                                    "id_karyawan": {
                                        "description": "ID karyawan yang menerima gaji",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "periode": {
                                        "description": "Periode gaji (Bulan/Tahun)",
                                        "type": "string",
                                        "example": "05/2025"
                                    },
                                    "gaji_pokok": {
                                        "description": "Gaji pokok karyawan",
                                        "type": "integer",
                                        "example": 5000000
                                    },
                                    "potongan": {
                                        "description": "Potongan gaji",
                                        "type": "integer",
                                        "example": 500000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gaji karyawan berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Gaji karyawan berhasil ditambahkan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/karyawan/gaji/update/{id_karyawan}": {
            "put": {
                "tags": [
                    "GajiKaryawan"
                ],
                "summary": "Update data gaji karyawan berdasarkan ID karyawan",
                "operationId": "UpdateGajiKaryawanByIdKaryawan",
                "parameters": [
                    {
                        "name": "id_karyawan",
                        "in": "path",
                        "description": "ID karyawan yang gajinya akan diupdate",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "periode",
                                    "gaji_pokok",
                                    "potongan"
                                ],
                                "properties": {
                                    "periode": {
                                        "description": "Periode gaji (Bulan/Tahun)",
                                        "type": "string",
                                        "example": "05/2025"
                                    },
                                    "gaji_pokok": {
                                        "description": "Gaji pokok karyawan",
                                        "type": "integer",
                                        "example": 5000000
                                    },
                                    "potongan": {
                                        "description": "Potongan gaji",
                                        "type": "integer",
                                        "example": 500000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gaji karyawan berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Gaji karyawan berhasil diupdate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data gaji tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data gaji tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/karyawan/gaji/delete/{id_karyawan}": {
            "delete": {
                "tags": [
                    "GajiKaryawan"
                ],
                "summary": "Hapus data gaji karyawan berdasarkan ID karyawan",
                "operationId": "DeleteGajiByKaryawanId",
                "parameters": [
                    {
                        "name": "id_karyawan",
                        "in": "path",
                        "description": "ID karyawan yang gajinya akan dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data gaji berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data gaji berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data gaji tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data gaji tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan saat menghapus data gaji",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Gagal menghapus data gaji: ..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "security": [
            {
                "bearerAuth": []
            }
        ]
    }
}